Refactor Common styled components
