+++ src/styled/Common/Types.ts
@@ -1,0 +1,28 @@
+export type overflowTypes = 'scroll' | 'auto' | 'hidden' | 'visible';
+export type backgrounds =
+    | 'dark1'
+    | 'dark2'
+    | 'dark3'
+    | 'dark4'
+    | 'title-gradient';
+
+export type textColors =
+    | 'text1'
+    | 'text2'
+    | 'text3'
+    | 'accent1'
+    | 'accent2'
+    | 'accent3'
+    | 'accent4'
+    | 'accent5'
+    | 'positive'
+    | 'negative'
+    | 'other-green'
+    | 'other-red'
+    | 'white'
+    | 'orange';
+
+export type Fonts = 'font-logo' | 'font-family' | 'roboto' | 'mono';
+
+export type FontSizes = 'header1' | 'header2' | 'header' | 'body' | 'mini';
+export type FontWeights = '100' | '200' | '300' | '400' | '500' | 'lighter';
