+++ src/styled/Common/index.ts
@@ -1,0 +1,14 @@
+import { FlexContainer, Container, GridContainer, Text } from './Common';
+import { Breakpoint } from './Breakpoints';
+import { ContainerStyles } from './Container';
+import { ScrollContainer } from './ScrollContainer';
+
+export {
+    ContainerStyles,
+    Breakpoint,
+    FlexContainer,
+    Container,
+    GridContainer,
+    ScrollContainer,
+    Text,
+};
