+++ src/styled/Components/Portfolio.ts
@@ -1,13 +1,9 @@
 import styled from 'styled-components/macro';
 import { motion } from 'framer-motion';
-import {
-    ContainerProps,
-    ContainerStyles,
-    BreakpointProps,
-    Breakpoint,
-    WrappedContainerStyles,
-    FlexContainer,
-} from '../Common';
+import { ContainerStyles, Breakpoint, FlexContainer } from '../Common';
+
+import { WrappedContainerStyles, ContainerProps } from '../Common/Container';
+import { BreakpointProps } from '../Common/Breakpoints';
 
 export const PortfolioTabsPortfolioTabsContainer = styled.div`
     width: 100%;
@@ -117,8 +113,7 @@ export const PortfolioControlContainer = styled.div`
 export const PortfolioMotionContainer = styled(motion.div)<
     ContainerProps & BreakpointProps
 >`
-    display: flex;
-    ${WrappedContainerStyles}
+    ${(props) => ContainerStyles({ ...props, ...{ display: 'flex' } })}
     ${Breakpoint}
 `;
